fix: wrong offset in test vector generator script
According to the latest draft, the represented lengths in a sframe header
(KLEN, LEN) are calculated by substracting 1 from the respective lenght in
bytes of the KID CTR values. Thus a length of 1..8 can be represented.
However this was not reflected in the generator script for test vectors,
which are wrongly created. I fixed this, recreated the test vectors
and could prove it working with our current reference implementation.